+++ qiskit/transpiler/passes/mapping/barrier_before_final_measurements.py
@@ -52,7 +52,7 @@ class BarrierBeforeFinalMeasurements(TransformationPass):
                            for final_op in final_ops)
 
         barrier_layer.apply_operation_back(Barrier(qubits=final_qubits))
-        new_barrier_id = barrier_layer.node_counter
+        new_barrier_id = barrier_layer._node_counter
 
         # Preserve order of final ops collected earlier from the original DAG.
         ordered_node_ids = [node_id for node_id in dag.node_nums_in_topological_order()
@@ -88,7 +88,7 @@ class BarrierBeforeFinalMeasurements(TransformationPass):
                 merge_barrier = Barrier(qubits=(our_qubits | their_qubits))
                 barrier_layer.apply_operation_front(merge_barrier)
 
-                merge_barrier_id = barrier_layer.node_counter
+                merge_barrier_id = barrier_layer._node_counter
                 our_ancestors = our_ancestors | their_ancestors
                 our_descendants = our_descendants | their_descendants